refactor: graphql resolve service
