+++ src/utils/DB/entities/DBMemberTypes.ts
@@ -5,7 +5,7 @@ export type MemberTypeEntity = {
   discount: number;
   monthPostsLimit: number;
 };
-type CreateMemberTypeDTO = MemberTypeEntity;
+export type CreateMemberTypeDTO = MemberTypeEntity;
 type ChangeMemberTypeDTO = Partial<Omit<MemberTypeEntity, 'id'>>;
 
 export default class DBMemberTypes extends DBEntity<
