+++ src/utils/DB/entities/DBProfiles.ts
@@ -12,8 +12,8 @@ export type ProfileEntity = {
   memberTypeId: string;
   userId: string;
 };
-type CreateProfileDTO = Omit<ProfileEntity, 'id'>;
-type ChangeProfileDTO = Partial<Omit<ProfileEntity, 'id' | 'userId'>>;
+export type CreateProfileDTO = Omit<ProfileEntity, 'id'>;
+export type ChangeProfileDTO = Partial<Omit<ProfileEntity, 'id' | 'userId'>>;
 
 export default class DBProfiles extends DBEntity<
   ProfileEntity,
