+++ src/utils/DB/entities/DBUsers.ts
@@ -8,8 +8,8 @@ export type UserEntity = {
   email: string;
   subscribedToUserIds: string[];
 };
-type CreateUserDTO = Omit<UserEntity, 'id' | 'subscribedToUserIds'>;
-type ChangeUserDTO = Partial<Omit<UserEntity, 'id'>>;
+export type CreateUserDTO = Omit<UserEntity, 'id' | 'subscribedToUserIds'>;
+export type ChangeUserDTO = Partial<Omit<UserEntity, 'id'>>;
 
 export default class DBUsers extends DBEntity<
   UserEntity,
